+++ src/components/POLine/POLineForm.js
@@ -56,6 +56,7 @@ import {
   ViewMetaData,
 } from '@folio/stripes/smart-components';
 
+import { AccordionInfoPopover } from '../../common';
 import {
   ENTITY_TYPE_PO_LINE,
   PO_LINE_CONFIG_NAME_PREFIX,
@@ -99,6 +100,7 @@ import { ItemForm } from './Item';
 import LocationForm from './Location/LocationForm';
 import { OngoingOrderForm } from './OngoingOrder';
 import { OtherForm } from './Other';
+import { PaymentTermsFormContainer } from './PaymentTerms';
 import { PhysicalForm } from './Physical';
 import { POLineDetailsForm } from './POLineDetails';
 import { VendorForm } from './Vendor';
@@ -147,6 +149,8 @@ function POLineForm({
   const fundDistribution = get(formValues, POL_FORM_FIELDS.fundDistribution, []);
   const lineLocations = get(formValues, POL_FORM_FIELDS.locations, []);
   const instanceId = formValues.instanceId;
+  const isOrderOngoing = isOngoing(order.orderType);
+  const isMultiYearPaymentActive = Boolean(formValues.multiYearPayment);
 
   const fundsMap = useMemo(() => keyBy(fundsRecords, 'id'), [fundsRecords]);
 
@@ -350,9 +354,17 @@ function POLineForm({
 
   const formErrors = form.getState()?.errors;
   const errors = useMemo(() => (
-    omitFieldArraysAsyncErrors(formErrors, [POL_FORM_FIELDS.fundDistribution])
+    omitFieldArraysAsyncErrors(formErrors, [
+      POL_FORM_FIELDS.fundDistribution,
+      `${POL_FORM_FIELDS.paymentTerms}.fiscalYearDistributions`,
+    ])
   ), [formErrors]);
   const errorAccordionStatus = useErrorAccordionStatus({ errors, fieldsMap: MAP_FIELD_ACCORDION });
+
+  const accordionsInitialStatus = useMemo(() => ({
+    ...INITIAL_SECTIONS,
+    [ACCORDION_ID.paymentTerms]: isMultiYearPaymentActive,
+  }), [isMultiYearPaymentActive]);
 
   const lineNumber = get(initialValues, 'poLineNumber', '');
   const firstMenu = getAddFirstMenu();
@@ -505,7 +517,7 @@ function POLineForm({
                         />
 
                         <AccordionSet
-                          initialStatus={INITIAL_SECTIONS}
+                          initialStatus={accordionsInitialStatus}
                           accordionStatus={{ ...status, ...errorAccordionStatus }}
                         >
                           <Accordion
@@ -560,13 +572,14 @@ function POLineForm({
                               />
                             </Accordion>
                           </IfFieldVisible>
-                          {isOngoing(order.orderType) && (
+                          {isOrderOngoing && (
                             <Accordion
                               label={<FormattedMessage id="ui-orders.line.accordion.ongoingOrder" />}
                               id={ACCORDION_ID.ongoingOrder}
                             >
                               <OngoingOrderForm
                                 hiddenFields={hiddenFields}
+                                order={order}
                               />
                             </Accordion>
                           )}
@@ -615,6 +628,28 @@ function POLineForm({
                               />
                             </Accordion>
                           </IfFieldVisible>
+
+                          {isOrderOngoing && (
+                            <IfFieldVisible
+                              visible={!hiddenFields?.paymentTerms}
+                              name={POL_FORM_FIELDS.paymentTerms}
+                            >
+                              <Accordion
+                                label={(
+                                  <>
+                                    <FormattedMessage id="ui-orders.line.accordion.paymentTerms" />
+                                    <AccordionInfoPopover content={<FormattedMessage id="ui-orders.line.accordion.paymentTerms.infoPopover" />} />
+                                  </>
+                                )}
+                                id={ACCORDION_ID.paymentTerms}
+                              >
+                                <PaymentTermsFormContainer
+                                  filterFunds={filterFunds}
+                                  order={order}
+                                />
+                              </Accordion>
+                            </IfFieldVisible>
+                          )}
 
                           <IfFieldVisible
                             visible={!hiddenFields?.locations}